Added: New pinia store for Ecom stores, facilities, carriers, shipment methods and facilitygroups
