+++ src/components/RoutingHistoryModal.vue
@@ -38,7 +38,8 @@
 
 <script setup lang="ts">
 import { useUserStore } from "@/store/userStore";
-import { useOrderRoutingStore } from "@/store/orderRoutingStore";
+import { orderRoutingStore } from "@/store/orderRoutingStore";
+import { productStore } from "@/store/productStore";
 import { translate } from "@common";
 import { IonButton, IonButtons, IonContent, IonHeader, IonIcon, IonItem, IonLabel, IonList, IonTitle, IonToolbar, modalController } from "@ionic/vue";
 import { checkmarkDoneOutline, closeOutline, warningOutline } from "ionicons/icons";
@@ -61,10 +62,10 @@ defineProps({
   }
 })
 
-const userProfile = computed(() => useUserStore().getUserProfile)
-const currentRoutingGroup: any = computed(() => useOrderRoutingStore().getCurrentRoutingGroup)
+const currentRoutingGroup: any = computed(() => orderRoutingStore().getCurrentRoutingGroup)
+const ecomStores = computed(() => productStore().ecomStores)
 
-const productStoreName = computed(() => userProfile.value.stores.find((store: any) => store.productStoreId === currentRoutingGroup.value.productStoreId)?.storeName || currentRoutingGroup.value.productStoreId)
+const productStoreName = computed(() => ecomStores.value.find((store: any) => store.productStoreId === currentRoutingGroup.value.productStoreId)?.storeName || currentRoutingGroup.value.productStoreId)
 
 function closeModal() {
   modalController.dismiss();
