+++ src/store/utilStore.ts
@@ -2,34 +2,19 @@ import { defineStore } from 'pinia'
 import { UtilService } from "@/services/UtilService"
 import { logger, commonUtil } from '@common'
 import { EnumerationAndType } from "@/types"
-import { useOrderRoutingStore } from './orderRoutingStore'
+import { orderRoutingStore } from './orderRoutingStore'
 
 export const useUtilStore = defineStore('util', {
   state: () => {
     return {
       enums: {} as any,
-      facilities: {} as any,
       categories: {} as any,
-      shippingMethods: {} as any,
-      facilityGroups: {} as any,
-      statuses: {} as any,
-      carriers: {} as any
+      statuses: {} as any
     }
   },
   getters: {
     getEnums(state) {
       return state.enums
-    },
-    getFacilities(state) {
-      return state.facilities
-    },
-    getVirtualFacilities(state) {
-      return Object.values(state.facilities).reduce((virtualFacilities: any, facility: any) => {
-        if(facility.parentTypeId === "VIRTUAL_FACILITY") {
-          virtualFacilities[facility.facilityId] = facility
-        }
-        return virtualFacilities;
-      }, {}) as any
     },
     getCatalogCategories(state) {
       return Object.values(state.categories).reduce((catalogCategories: any, category: any) => {
@@ -39,25 +24,8 @@ export const useUtilStore = defineStore('util', {
         return catalogCategories;
       }, {})
     },
-    getPhysicalFacilities(state) {
-      return Object.values(state.facilities).reduce((virtualFacilities: any, facility: any) => {
-        if(facility.parentTypeId !== "VIRTUAL_FACILITY") {
-          virtualFacilities[facility.facilityId] = facility
-        }
-        return virtualFacilities;
-      }, {})
-    },
-    getShippingMethods(state) {
-      return state.shippingMethods
-    },
-    getFacilityGroups(state) {
-      return state.facilityGroups
-    },
     getStatusDesc: (state) => (id: any) => {
       return state.statuses[id]?.description ? state.statuses[id]?.description : id
-    },
-    getCarriers(state) {
-      return state.carriers
     }
   },
   actions: {
@@ -149,29 +117,7 @@ export const useUtilStore = defineStore('util', {
       }
       this.enums = enums;
     },
-    async fetchFacilities() {
-      let facilities = JSON.parse(JSON.stringify(this.facilities))
-  
-      if(Object.keys(facilities).length) {
-        return;
-      }
-  
-      const payload = { pageSize: 500 }
-  
-      try {
-        const resp = await UtilService.fetchFacilities(payload);
-  
-        if(!commonUtil.hasError(resp) && resp.data.length) {
-          facilities = resp.data.reduce((facilities: any, facility: any) => {
-            facilities[facility.facilityId] = facility
-            return facilities
-          }, {})
-        }
-      } catch(err) {
-        logger.error(err)
-      }
-      this.facilities = facilities;
-    },
+
     async fetchCategories() {
       let categories = JSON.parse(JSON.stringify(this.categories))
       if(Object.keys(categories).length) return;
@@ -179,7 +125,7 @@ export const useUtilStore = defineStore('util', {
       let pageIndex = 0
       const pageSize = 500
       const basePayload = {
-        productStoreId: useOrderRoutingStore().currentGroup.productStoreId,
+        productStoreId: orderRoutingStore().currentGroup.productStoreId,
         pageSize
       }
   
@@ -203,50 +149,7 @@ export const useUtilStore = defineStore('util', {
       }
       this.categories = categories;
     },
-    async fetchShippingMethods() {
-      let shippingMethods = JSON.parse(JSON.stringify(this.shippingMethods))
-      if(Object.keys(shippingMethods).length) return;
-  
-      const payload = {
-        productStoreId: useOrderRoutingStore().currentGroup.productStoreId,
-        pageSize: 200
-      }
-  
-      try {
-        const resp = await UtilService.fetchShippingMethods(payload);
-        if(!commonUtil.hasError(resp) && resp.data.length) {
-          shippingMethods = resp.data.reduce((shippingMethods: any, shippingMethod: any) => {
-            shippingMethods[shippingMethod.shipmentMethodTypeId] = shippingMethod
-            return shippingMethods
-          }, {})
-        }
-      } catch(err) {
-        logger.error(err)
-      }
-      this.shippingMethods = shippingMethods;
-    },
-    async fetchFacilityGroups() {
-      let facilityGroups = JSON.parse(JSON.stringify(this.facilityGroups))
-      if(Object.keys(facilityGroups).length) return;
-  
-      const payload = {
-        productStoreId: useOrderRoutingStore().currentGroup.productStoreId,
-        pageSize: 200
-      }
-  
-      try {
-        const resp = await UtilService.fetchFacilityGroups(payload);
-        if(!commonUtil.hasError(resp) && resp.data.length) {
-          facilityGroups = resp.data.reduce((facilityGroups: any, facilityGroup: any) => {
-            facilityGroups[facilityGroup.facilityGroupId] = facilityGroup
-            return facilityGroups
-          }, {})
-        }
-      } catch(err) {
-        logger.error(err)
-      }
-      this.facilityGroups = facilityGroups;
-    },
+
     async fetchStatusInformation() {
       let statuses = JSON.parse(JSON.stringify(this.statuses))
       if(Object.keys(statuses).length) return;
@@ -266,84 +169,11 @@ export const useUtilStore = defineStore('util', {
       }
       this.statuses = statuses;
     },
-    async fetchCarrierInformation(carrierIds: Array<any>) {
-      let carriers = JSON.parse(JSON.stringify(this.carriers))
-      const carrierPartyIds = carrierIds.filter((id: any) => !carriers[id])
-  
-      if(!carrierPartyIds.length) return;
-  
-      const payload = {
-        inputFields: {
-          partyId: carrierIds,
-          partyId_op: "in"
-        },
-        distinct: "Y",
-        viewSize: carrierIds.length,
-        entityName: "PartyNameView",
-      }
-  
-      try {
-        const resp = await UtilService.getCarrierInformation(payload);
-        if(!commonUtil.hasError(resp) && resp.data.docs?.length) {
-          carriers = resp.data.docs.reduce((carriers: any, carrier: any) => {
-            carriers[carrier.partyId] = {
-              name: carrier.groupName
-            }
-            return carriers
-          }, carriers)
-        }
-      } catch(err) {
-        logger.error(err)
-      }
-  
-      const deliveryDaysPayload = {
-        inputFields: {
-          partyId: carrierIds,
-          partyId_op: "in",
-          roleTypeId: "CARRIER",
-          deliveryDays_op: "not-empty"
-        },
-        distinct: "Y",
-        viewSize: 200,
-        entityName: "CarrierShipmentMethod",
-      }
-  
-      try {
-        const resp = await UtilService.getCarrierDeliveryDays(deliveryDaysPayload);
-        if(!commonUtil.hasError(resp) && resp.data.docs?.length) {
-          carriers = resp.data.docs.reduce((carriers: any, carrier: any) => {
-            if(carriers[carrier.partyId]["deliveryDays"]) {
-              carriers[carrier.partyId]["deliveryDays"] = {
-                ...carriers[carrier.partyId]["deliveryDays"],
-                [carrier.shipmentMethodTypeId]: carrier.deliveryDays
-              }
-            } else {
-              carriers[carrier.partyId]["deliveryDays"] = {
-                [carrier.shipmentMethodTypeId]: carrier.deliveryDays
-              }
-            }
-            return carriers
-          }, carriers)
-        }
-      } catch(err) {
-        logger.error(err)
-      }
-      this.carriers = carriers;
-    },
+
     async clearUtilState() {
         this.enums = {};
-        this.facilities = {};
         this.categories = {};
-        this.shippingMethods = {};
-        this.facilityGroups = {};
         this.statuses = {};
-        this.carriers = {};
-    },
-    async updateShippingMethods(payload: any) {
-      this.shippingMethods = payload;
-    },
-    async updateFacillityGroups(payload: any) {
-      this.facilityGroups = payload;
     },
     async updateProductCategories(payload: any) {
       this.categories = payload;
